Move tasks to model package
